Delete Question Completed

+++ MockExamSpringBoot/src/main/java/com/lxisoft/mockExamSpringBoot/dao/MockDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lxisoft.mockExamSpringBoot.entity.MockEntity;
 
-public interface MockDao extends JpaRepository<MockEntity, Long> {
+public interface MockDao extends JpaRepository<MockEntity, Integer> {
 
 }
